Add history db sqlite format support.
gpbackup functionality of the `--gpbackup.history-file` flag has been expanded. Now it can be specified as `gpbackup_history.db` (sqllite format) or `gpbackup_history.yaml` (yaml format). If the file extension does not match, an error is returned and the exporter does not start.

In the future, the yaml format will be abandoned (see how gpbackup develops).
So this check will become irrelevant.